Change resource-related fields to short for compat with 32-bit compilers.
When cross-compiling the utilities are 32-bit so need unsigned short
instead of unsigned int for things that need to be 16 bits.


git-svn-id: https://svn.code.sf.net/p/freedos/svn/freecom/trunk@1749 6ac86273-5f31-0410-b378-82cca8765d1b

+++ include/infores.h
@@ -62,11 +62,11 @@ typedef struct {
 
 #include <algnbyte.h>
 struct EXE_header {
-	unsigned sig, lastFill, numBlocks, numReloc;
-	unsigned header;
-	unsigned extraMin, extraMax;
-	unsigned fSS, fSP, checksum, fIP, fCS;
-	unsigned offReloc, ovrlyNum;
+	unsigned short sig, lastFill, numBlocks, numReloc;
+	unsigned short header;
+	unsigned short extraMin, extraMax;
+	unsigned short fSS, fSP, checksum, fIP, fCS;
+	unsigned short offReloc, ovrlyNum;
 };
 #include <algndflt.h>
